fix(alearm246): move project service to google-manager service
