+++ api/project/controllers/project.js
@@ -4,8 +4,6 @@
  * Read the documentation (https://strapi.io/documentation/developer-docs/latest/development/backend-customization.html#core-controllers)
  * to customize this controller
  */
-const {sanitizeEntity} = require('strapi-utils');
-
 const { parseMultipartData, sanitizeEntity } = require('strapi-utils');
